feat: 6.1, more refactoring
